Support polyglot BUILD file
**Problem**
Currently it's not possible to mix languages within a package.

**Solution**
1. This implements `--append` flag to allow appending contents
   to BUILD.bazel.
2. This also adds `include_file_extension` configuration to
   allow `py_test` target to keep the name of the file
   as the target name.

+++ crates/driver/src/print_build.rs
@@ -1,6 +1,6 @@
 use std::{
     collections::{HashMap, HashSet},
-    path::PathBuf,
+    path::{PathBuf, Path},
     sync::Arc,
 };
 
@@ -18,7 +18,7 @@ use ignore::WalkBuilder;
 use rustpython_parser::ast;
 
 use futures::Future;
-use tokio::sync::Semaphore;
+use tokio::{sync::Semaphore, io::AsyncWriteExt};
 
 lazy_static::lazy_static! {
     static ref BUILD_BAZEL: std::ffi::OsString = std::ffi::OsString::from("BUILD.bazel");
@@ -259,33 +259,7 @@ where
     for graph_node in graph_nodes {
         let node_file = opt.working_directory.join(&graph_node.node_label);
         let node_file_name = to_file_name(&node_file);
-        let target_name = if !opt.no_aggregate_source {
-            base_name.clone()
-        } else {
-            node_file_name.replace(".", "_")
-        };
         let mut extra_kv_pairs: HashMap<String, Vec<String>> = HashMap::default();
-
-        fn add_non_empty(
-            opt: &'static Opt,
-            key: &str,
-            labels: &Vec<String>,
-            extra_kv_pairs: &mut HashMap<String, Vec<String>>,
-        ) {
-            if !labels.is_empty() {
-                let vals = labels.iter().map(|e| to_label(&opt, e)).collect();
-                extra_kv_pairs.insert(key.to_string(), vals);
-            }
-        }
-
-        add_non_empty(opt, "deps", &graph_node.dependencies, &mut extra_kv_pairs);
-        add_non_empty(
-            opt,
-            "runtime_deps",
-            &graph_node.runtime_dependencies,
-            &mut extra_kv_pairs,
-        );
-
         let (build_config, use_rglob) = if path_type == "test" {
             (&module_config.build_config.test, true)
         } else {
@@ -300,6 +274,35 @@ where
                 graph_node
             ));
         };
+
+        let include_file_extension = build_config.include_file_extension;
+        let target_name = if !opt.no_aggregate_source {
+            base_name.clone()
+        } else {
+            to_name_from_file_name(&node_file_name, include_file_extension)?
+        }; 
+
+        fn add_non_empty(
+            opt: &'static Opt,
+            key: &str,
+            labels: &Vec<String>,
+            extra_kv_pairs: &mut HashMap<String, Vec<String>>,
+            include_file_extension: bool,
+        ) {
+            if !labels.is_empty() {
+                let vals = labels.iter().map(|e| to_label(&opt, e, include_file_extension)).collect();
+                extra_kv_pairs.insert(key.to_string(), vals);
+            }
+        }
+
+        add_non_empty(opt, "deps", &graph_node.dependencies, &mut extra_kv_pairs, include_file_extension);
+        add_non_empty(
+            opt,
+            "runtime_deps",
+            &graph_node.runtime_dependencies,
+            &mut extra_kv_pairs,
+            include_file_extension,
+        );
 
         for (k, lst) in build_config.extra_key_to_list.iter() {
             append_key_values(&mut extra_kv_pairs, &k, &lst);
@@ -501,7 +504,7 @@ where
         );
     } // end for graph_nodes
 
-    fn to_label(opt: &'static Opt, entry: &String) -> String {
+    fn to_label(opt: &'static Opt, entry: &String, include_file_extension: bool) -> String {
         if entry.starts_with('@') {
             entry.clone()
         } else {
@@ -510,8 +513,19 @@ where
             } else {
                 let directory = to_directory(entry.to_string());
                 let full_file = opt.working_directory.join(&entry);
-                let name = to_file_name(&full_file).replace(".", "_");
+                let name = to_name_from_file_name(&to_file_name(&full_file), include_file_extension).unwrap();
                 format!("//{}:{}", directory, name)
+            }
+        }
+    }
+
+    fn to_name_from_file_name(file_name: &String, include_file_extension: bool) -> Result<String> {
+        if include_file_extension {
+            Ok(file_name.replace(".", "_"))
+        } else {
+            match Path::new(&file_name).file_stem() {
+                Some(s) => Ok(s.to_str().unwrap().to_string()),
+                None => Err(anyhow!("can't get file_stem of {}", file_name))
             }
         }
     }
@@ -743,9 +757,25 @@ async fn print_file(
     )
     .await?;
     let handle = concurrent_io_operations.acquire().await?;
-    tokio::fs::write(&target_file, t.emit_build_file()?)
-        .await
-        .with_context(|| format!("Attempting to write file data to {:?}", target_file))?;
+    if opt.append {
+        let mut file = tokio::fs::OpenOptions::new()
+            .write(true)
+            .create(true)
+            .append(true)
+            .open(&target_file).await?;
+        file.write(t.emit_build_file()?.as_bytes())
+            .await
+            .with_context(|| format!("Attempting to write file data to {:?}", target_file))?;
+    } else {
+        let mut file = tokio::fs::OpenOptions::new()
+            .write(true)
+            .create(true)
+            .truncate(true)
+            .open(&target_file).await?;
+        file.write_all(t.emit_build_file()?.as_bytes())
+            .await
+            .with_context(|| format!("Attempting to write file data to {:?}", target_file))?;
+    }
     drop(handle);
 
     Ok(emitted_files)
@@ -837,9 +867,11 @@ pub async fn print_build(
     }
 
     // These files are old and not updated..
-    for f in current_files {
-        println!("Deleting no longer used build file of: {:?}", f);
-        std::fs::remove_file(&f)?;
+    if !opt.append {
+        for f in current_files {
+            println!("Deleting no longer used build file of: {:?}", f);
+            std::fs::remove_file(&f)?;
+        }
     }
 
     Ok(())
@@ -853,13 +885,14 @@ mod tests {
     use crate::Commands::PrintBuild;
     use std::collections::BTreeMap;
 
-    fn example_opt(no_aggregate_source: bool) -> Opt {
+    fn example_opt(no_aggregate_source: bool, append: bool) -> Opt {
         Opt {
             input_path: PathBuf::new(),
             working_directory: PathBuf::new(),
             concurrent_io_operations: 8,
             cache_path: PathBuf::new(),
             no_aggregate_source: no_aggregate_source,
+            append: append,
             command: PrintBuild(PrintBuildArgs {
                 graph_data: PathBuf::new(),
             }),
@@ -879,6 +912,7 @@ mod tests {
                                 load_value: "proto_library".to_string(),
                             }],
                             function_name: "proto_library".to_string(),
+                            include_file_extension: true,
                             extra_key_to_list: HashMap::default(),
                             extra_key_to_value: HashMap::default(),
                         }),
@@ -908,6 +942,7 @@ mod tests {
                                 load_value: "proto_library".to_string(),
                             }],
                             function_name: "proto_library".to_string(),
+                            include_file_extension: true,
                             extra_key_to_list: HashMap::default(),
                             extra_key_to_value: HashMap::default(),
                         }),
@@ -919,6 +954,7 @@ mod tests {
                                 GrpBuildConfig {
                                     headers: vec![],
                                     function_name: "java_proto_library".to_string(),
+                                    include_file_extension: true,
                                     extra_key_to_list: HashMap::from([(
                                         "deps".to_string(),
                                         vec![":${name}".to_string()],
@@ -935,6 +971,7 @@ mod tests {
                                         load_value: "py_proto_library".to_string(),
                                     }],
                                     function_name: "py_proto_library".to_string(),
+                                    include_file_extension: true,
                                     extra_key_to_list: HashMap::from([
                                         ("srcs".to_string(), vec!["${srcs}".to_string()]),
                                         ("deps".to_string(), vec!["${deps}_py".to_string()]),
@@ -977,6 +1014,7 @@ proto_library(
     visibility=['//visibility:public']
 )
         "#,
+            false,
             false,
         )
         .await
@@ -1042,6 +1080,7 @@ py_proto_library(
 )
         "#,
             true,
+            false,
         )
         .await
     }
@@ -1053,9 +1092,10 @@ py_proto_library(
         expected_target_count: usize,
         expected_build_file: &str,
         no_aggregate_source: bool,
+        append: bool,
     ) -> Result<(), Box<dyn std::error::Error>> {
         let mut emitted_files: Vec<PathBuf> = Vec::default();
-        let opt = Box::leak(Box::new(example_opt(no_aggregate_source)));
+        let opt = Box::leak(Box::new(example_opt(no_aggregate_source, append)));
         let boxed_project_conf = Box::leak(Box::new(project_conf));
         let target_entries = generate_targets(
             opt,
